Futher clean up
