+++ script/standard/test_alerts.rb
@@ -9,14 +9,14 @@ require_rel '../../test_support'
 
 asof_date = Date.new(2023, 7, 19)
 # schools = ['t*']
-schools = ['n*']
+schools = ['s*']
 
 overrides = {
   schools:  schools,
   cache_school: false,
   alerts:   { alerts: nil, control: { asof_date: asof_date} },
   alerts:   { alerts: [
-    AlertElectricityPeakKWVersusBenchmark
+    AlertSeasonalHeatingSchoolDays
 #    AlertThermostaticControl
 #    AlertEnergyAnnualVersusBenchmark,
 #    AlertSchoolWeekComparisonGas,
